move servers to internal, we're not blowing any minds here
