+++ cmd/hauler/cli/serve/registry.go
@@ -11,7 +11,7 @@ import (
 	"github.com/distribution/distribution/v3/version"
 	"github.com/spf13/cobra"
 
-	"github.com/rancherfederal/hauler/pkg/server"
+	"github.com/rancherfederal/hauler/internal/server"
 )
 
 type RegistryOpts struct {
