+++ internal/server/registry.go
